refactor common code
